Extract logic into hook https://github.com/CruGlobal/mpdx-react/pull/1737\#discussion_r3126255662
